Renamed MarkovChainKeyword to KeywordMarkovChain and added a comment explaining what it does and what it is used for.
